+++ packages/Zooniverse-React-Components/src/components/layout/admin-layout-indicator.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function AdminLayoutIndicator() {
+  return (<div className="admin-layout-indicator" title="Admin mode on!" />);
+}