+++ packages/Zooniverse-React-Components/test/setup.js
@@ -2,24 +2,27 @@
 /* eslint import/no-extraneous-dependencies: ["error", { "devDependencies": true  }] */
 import React from 'react';
 import chai from 'chai';
-import { jsdom } from 'jsdom';
+import dirtyChai from 'dirty-chai';
+import { JSDOM } from 'jsdom';
 
-// Set up fake DOM for use by Enzyme's mount() method.
-const exposedProperties = ['window', 'navigator', 'document'];
 
+chai.use(dirtyChai);
 global.React = React;
 global.expect = chai.expect;
 
-global.document = jsdom('');
-global.window = document.defaultView;
+const jsdom = new JSDOM('<!doctype html><html><body></body></html>');
+const { window } = jsdom;
 
-Object.keys(document.defaultView).forEach((property) => {
-  if (typeof global[property] === 'undefined') {
-    exposedProperties.push(property);
-    global[property] = document.defaultView[property];
-  }
-});
+function copyProps(src, target) {
+  const props = Object.getOwnPropertyNames(src)
+    .filter(prop => typeof target[prop] === 'undefined')
+    .map(prop => Object.getOwnPropertyDescriptor(src, prop));
+  Object.defineProperties(target, props);
+}
 
+global.window = window;
+global.document = window.document;
 global.navigator = {
   userAgent: 'node.js'
 };
+copyProps(window, global);